+++ Sources/BrazeUI/InAppMessageUI/Views/InAppMessageUIFullView.swift
@@ -55,17 +55,17 @@ extension BrazeInAppMessageUI {
       public var shadow: Shadow? = Shadow.inAppMessage
 
       /// The minimum width (used when displayed as modal).
-      public var minWidth = 320.0
+      public var minWidth: ViewDimension = 320.0
 
       /// The maximum width (used when displayed as modal).
       ///
       /// The maximum width is swapped with the maximum height for large UIs (e.g. iPad).
-      public var maxWidth = 450.0
+      public var maxWidth: ViewDimension = .init(regular: 450.0, large: 720.0)
 
       /// The maximum height (used when displayed as modal).
       ///
       /// The maximum height is swapped with the maximum width for large UIs (e.g. iPad).
-      public var maxHeight = 720.0
+      public var maxHeight: ViewDimension = .init(regular: 720.0, large: 450.0)
 
       /// Specify whether the full in-app message view can be dismissed from a tap with the
       /// view's background.
@@ -132,7 +132,7 @@ extension BrazeInAppMessageUI {
     }
 
     var modalMargin: UIEdgeInsets
-    var modalMaxHeight: Double
+    var modalMaxHeight: ViewDimension
     var modalCornerRadius: Double
     var modalPaddingBottom: Double